remove linefeeds from sql strings in viewer.jsons

+++ packages/migrate-viewer.ts
@@ -1,12 +1,16 @@
+// to run this script
+// node --experimental-specifier-resolution node --loader ts-node/esm ./migrate-viewer.ts
 import { Experience } from './lib/'
-import { writeFileSync, mkdirSync } from 'fs'
+import { writeFileSync, mkdirSync, readFileSync, existsSync } from 'fs'
 import * as packages from './packages'
 
 import camelCase from 'lodash/camelCase'
 import isEqual from 'lodash/isEqual'
-import { ViewerOptions } from './lib/types'
-
+import { ViewBlock, ViewerOptions } from './lib/types'
 const outputDir = 'viewer-opts'
+const iconUrlPrefix =
+  'https://raw.githubusercontent.com/hestiaAI/hestialabs-experiences/master/packages/lib/icons'
+
 function test([name, experience]: [
   name: string,
   experience: Experience
@@ -16,15 +20,47 @@ function test([name, experience]: [
     viewerOptions
     // viewerOptions: { viewBlocks, url }
   } = experience
-  viewerOptions.icon = 'TODO'
+  const viewerOptionsFileName = `${experience.name}-viewer.json`
+  const migratedViewerOptionsPath = `packages/experiences/${experience.name}/src/${viewerOptionsFileName}`
+  if (existsSync(migratedViewerOptionsPath)) {
+    console.log(`[${experience.name}] DONE: ${migratedViewerOptionsPath}`)
+    return
+  }
+  viewerOptions.viewBlocks.forEach(replaceSqlLineEndings)
   const unserializable = diffWithSerialized(viewerOptions)
   if (unserializable) {
-    console.log(`[${name}] viewerOptions are not serializable`)
-    console.log(JSON.stringify(unserializable, stringifyReplacer, 2))
+    console.log(`[${experience.name}] BAD ${formatDiff(unserializable)}`)
+    // console.log(JSON.stringify(unserializable, stringifyReplacer, 2))
+    // console.log(
+    //   ' ',
+    //   JSON.stringify(unserializable, stringifyReplacer).substring(0, 80)
+    // )
+  } else {
+    const fileName = `${outputDir}/${viewerOptionsFileName}`
+    const iconFile = experienceIcons[experience.name]
+    viewerOptions.icon = `${iconUrlPrefix}/${iconFile}`
+    writeFileSync(fileName, JSON.stringify(viewerOptions, null, 2))
+    console.log(`[${experience.name}] OK ${viewerOptionsFileName}`)
   }
-  const fileName = `${outputDir}/${experience.name}-viewer.json`
-  writeFileSync(fileName, JSON.stringify(viewerOptions, null, 2))
-  console.log('wrote', fileName)
+}
+
+function formatDiff(diff: any) {
+  const { viewBlocks } = diff
+  if (!viewBlocks) {
+    return ''
+  }
+  return Object.keys(viewBlocks)
+    .map(k => {
+      const b = viewBlocks[k]
+      return `${k}${b.customPipeline ? 'c' : ''}${b.postprocessor ? 'p' : ''}`
+    })
+    .join(' ')
+}
+
+function replaceSqlLineEndings(viewBlock: ViewBlock) {
+  if (viewBlock.sql) {
+    viewBlock.sql = viewBlock.sql.replace(/\r\n */g, ' ')
+  }
 }
 
 function saveUberDriverKeplerConfigs(viewerOptions: ViewerOptions) {
@@ -65,25 +101,17 @@ function diffObjects(
         const difference = diffObjects(val1, val2)
         // console.log('value different for key', key)
         differences[key] = difference
+        // differences[key] = 'value changed'
       }
     } else {
-      // console.log('key missing in 1:', key)
-      // Key only exists in the first object
-      differences[key] = {
-        oldValue: obj1[key],
-        newValue: undefined
-      }
+      differences[key] = 'missing'
     }
   }
 
   // Check for keys in the second object that are not in the first object
   for (const key in obj2) {
     if (!(key in obj1)) {
-      // console.log('key missing in 2:', key)
-      differences[key] = {
-        oldValue: undefined,
-        newValue: obj2[key]
-      }
+      differences[key] = 'appeared'
     }
   }
 
@@ -123,6 +151,12 @@ function removeUndefinedKeys(object: any) {
   }
   return object
 }
+
+// Change this typescript code so that it doesn't produce this error
+// Argument of type 'Buffer' is not assignable to parameter of type 'string'. (lsp)
+const experienceIconsBuf = readFileSync('./experience-icons.json')
+const experienceIcons = JSON.parse(experienceIconsBuf.toString())
+mkdirSync(outputDir, { recursive: true } as any)
 
 if (process.argv.length > 2) {
   // test one or more packages
